Moved CameraData code to camera.cc/.h
