+++ source/blender/render/hydra/viewportEngine.cc
@@ -13,28 +13,13 @@
 #include "glog/logging.h"
 
 #include "viewportEngine.h"
+#include "sceneDelegate/camera.h"
 #include "utils.h"
 
 using namespace std;
 using namespace pxr;
 
 namespace blender::render::hydra {
-
-struct CameraData {
-  CameraData(Object *camera_obj, GfVec2i res, GfVec4f tile);
-  CameraData(BL::Context &b_context);
-
-  GfCamera export_gf(GfVec4f tile);
-
-  int mode;
-  GfRange1f clip_range;
-  float focal_length;
-  GfVec2f sensor_size;
-  GfMatrix4d transform;
-  GfVec2f lens_shift;
-  GfVec2f ortho_size;
-  tuple<float, float, int> dof_data;
-};
 
 struct ViewSettings {
   ViewSettings(BL::Context &b_context);
@@ -50,268 +35,6 @@ struct ViewSettings {
   int screen_height;
   GfVec4i border;
 };
-
-CameraData::CameraData(Object *camera_obj, GfVec2i res, GfVec4f tile)
-{
-  Camera *camera = (Camera *)camera_obj->data;
-  
-  float t_pos[2] = {tile[0], tile[1]};
-  float t_size[2] = {tile[2], tile[3]};
-  transform = gf_matrix_from_transform(camera_obj->object_to_world);
-  clip_range = GfRange1f(camera->clip_start, camera->clip_end);
-  mode = camera->type;
-
-  if (camera->dof.flag & CAM_DOF_ENABLED)
-  {
-    float focus_distance;
-    if (!camera->dof.focus_object) {
-      focus_distance = camera->dof.focus_distance;
-    } 
-    else {
-      GfVec3f obj_pos(camera->dof.focus_object->object_to_world[0][3],
-                      camera->dof.focus_object->object_to_world[1][3],
-                      camera->dof.focus_object->object_to_world[2][3]);
-      GfVec3f cam_pos(transform[0][3],
-                      transform[1][3],
-                      transform[2][3]);
-      focus_distance = (obj_pos - cam_pos).GetLength();
-    }
-
-    dof_data = tuple(max(focus_distance, 0.001f),
-                     camera->dof.aperture_fstop,
-                     camera->dof.aperture_blades);
-  }
-
-  float ratio = (float)res[0] / res[1];
-
-  switch (camera->sensor_fit) {
-    case CAMERA_SENSOR_FIT_VERT:
-      lens_shift = GfVec2f(camera->shiftx / ratio, camera->shifty);
-      break;
-    case CAMERA_SENSOR_FIT_HOR:
-      lens_shift = GfVec2f(camera->shiftx, camera->shifty * ratio);
-      break;
-    case CAMERA_SENSOR_FIT_AUTO:
-      if (ratio > 1.0f) {
-        lens_shift = GfVec2f(camera->shiftx, camera->shifty * ratio);
-      }
-      else {
-        lens_shift = GfVec2f(camera->shiftx / ratio, camera->shifty);
-      }
-      break;
-    default:
-      lens_shift = GfVec2f(camera->shiftx, camera->shifty);
-      break;
-  }
-
-  lens_shift = GfVec2f(lens_shift[0] / t_size[0] + (t_pos[0] + t_size[0] * 0.5 - 0.5) / t_size[0],
-                       lens_shift[1] / t_size[1] + (t_pos[1] + t_size[1] * 0.5 - 0.5) / t_size[1]);
-
-  switch (camera->type) {
-    case CAM_PERSP:
-      focal_length = camera->lens;
-
-      switch (camera->sensor_fit) {
-        case CAMERA_SENSOR_FIT_VERT:
-          sensor_size = GfVec2f(camera->sensor_y * ratio, camera->sensor_y);
-          break;
-        case CAMERA_SENSOR_FIT_HOR:
-          sensor_size = GfVec2f(camera->sensor_x, camera->sensor_x / ratio);
-          break;
-        case CAMERA_SENSOR_FIT_AUTO:
-          if (ratio > 1.0f) {
-            sensor_size = GfVec2f(camera->sensor_x, camera->sensor_x / ratio);
-          }
-          else {
-            sensor_size = GfVec2f(camera->sensor_x * ratio, camera->sensor_x);
-          }
-          break;
-        default:
-          sensor_size = GfVec2f(camera->sensor_x, camera->sensor_y);
-          break;
-      }
-      sensor_size = GfVec2f(sensor_size[0] * t_size[0], sensor_size[1] * t_size[1]);
-      break;
-
-    case CAM_ORTHO:
-      focal_length = 0.0f;
-      switch (camera->sensor_fit) {
-        case CAMERA_SENSOR_FIT_VERT:
-          ortho_size = GfVec2f(camera->ortho_scale * ratio, camera->ortho_scale);
-          break;
-        case CAMERA_SENSOR_FIT_HOR:
-          ortho_size = GfVec2f(camera->ortho_scale, camera->ortho_scale / ratio);
-          break;
-        case CAMERA_SENSOR_FIT_AUTO:
-          if (ratio > 1.0f) {
-            ortho_size = GfVec2f(camera->ortho_scale, camera->ortho_scale / ratio);
-          }
-          else {
-            ortho_size = GfVec2f(camera->ortho_scale * ratio, camera->ortho_scale);
-          }
-          break;
-        default:
-          ortho_size = GfVec2f(camera->ortho_scale, camera->ortho_scale);
-          break;
-      }
-      ortho_size = GfVec2f(ortho_size[0] * t_size[0], ortho_size[1] * t_size[1]);
-      break;
-
-    case CAM_PANO:
-      /* TODO: Recheck parameters for PANO camera */
-      focal_length = camera->lens;
-
-      switch (camera->sensor_fit) {
-        case CAMERA_SENSOR_FIT_VERT:
-          sensor_size = GfVec2f(camera->sensor_y * ratio, camera->sensor_y);
-          break;
-        case CAMERA_SENSOR_FIT_HOR:
-          sensor_size = GfVec2f(camera->sensor_x, camera->sensor_x / ratio);
-          break;
-        case CAMERA_SENSOR_FIT_AUTO:
-          if (ratio > 1.0f) {
-            sensor_size = GfVec2f(camera->sensor_x, camera->sensor_x / ratio);
-          }
-          else {
-            sensor_size = GfVec2f(camera->sensor_x * ratio, camera->sensor_x);
-          }
-          break;
-        default:
-          sensor_size = GfVec2f(camera->sensor_x, camera->sensor_y);
-          break;
-      }
-      sensor_size = GfVec2f(sensor_size[0] * t_size[0], sensor_size[1] * t_size[1]);
-
-    default:
-      focal_length = camera->lens;
-      sensor_size = GfVec2f(camera->sensor_y * ratio, camera->sensor_y);
-  }
-}
-
-CameraData::CameraData(BL::Context &b_context)
-{
-  // this constant was found experimentally, didn't find such option in
-  // context.space_data or context.region_data
-  float VIEWPORT_SENSOR_SIZE = 72.0;
-
-  BL::SpaceView3D space_data = (BL::SpaceView3D)b_context.space_data();
-
-  GfVec2i res(b_context.region().width(), b_context.region().height());
-  float ratio = (float)res[0] / res[1];
-  transform = gf_matrix_from_transform((float(*)[4])b_context.region_data().view_matrix().data).GetInverse();
-
-  switch (b_context.region_data().view_perspective()) {
-    case BL::RegionView3D::view_perspective_PERSP: {
-      mode = CAM_PERSP;
-      clip_range = GfRange1f(space_data.clip_start(), space_data.clip_end());
-      lens_shift = GfVec2f(0.0, 0.0);
-      focal_length = space_data.lens();
-
-      if (ratio > 1.0) {
-        sensor_size = GfVec2f(VIEWPORT_SENSOR_SIZE, VIEWPORT_SENSOR_SIZE / ratio);
-      }
-      else {
-        sensor_size = GfVec2f(VIEWPORT_SENSOR_SIZE * ratio, VIEWPORT_SENSOR_SIZE);
-      }
-      break;
-    }
-
-    case BL::RegionView3D::view_perspective_ORTHO: {
-      mode = CAM_ORTHO;
-      lens_shift = GfVec2f(0.0f, 0.0f);
-
-      float o_size = b_context.region_data().view_distance() * VIEWPORT_SENSOR_SIZE / space_data.lens();
-      float o_depth = space_data.clip_end();
-
-      clip_range = GfRange1f(-o_depth * 0.5, o_depth * 0.5);
-
-      if (ratio > 1.0f) {
-        ortho_size = GfVec2f(o_size, o_size / ratio);
-      } else {
-        ortho_size = GfVec2f(o_size * ratio, o_size);
-      }
-      break;
-    }
-
-    case BL::RegionView3D::view_perspective_CAMERA: {
-      BL::Object camera_obj = space_data.camera();
-
-      GfMatrix4d mat = transform;
-      *this = CameraData((Object *)camera_obj.ptr.data, res, GfVec4f(0, 0, 1, 1));
-      transform = mat;
-
-      // This formula was taken from previous plugin with corresponded comment
-      // See blender/intern/cycles/blender/blender_camera.cpp:blender_camera_from_view (look for 1.41421f)
-      float zoom = 4.0 / pow((pow(2.0, 0.5) + b_context.region_data().view_camera_zoom() / 50.0), 2);
-
-      // Updating l_shift due to viewport zoom and view_camera_offset
-      // view_camera_offset should be multiplied by 2
-      lens_shift = GfVec2f((lens_shift[0] + b_context.region_data().view_camera_offset()[0] * 2) / zoom,
-                           (lens_shift[1] + b_context.region_data().view_camera_offset()[1] * 2) / zoom);
-
-      if (mode == BL::Camera::type_ORTHO) {
-        ortho_size *= zoom;
-      }
-      else {
-        sensor_size *= zoom;
-      }
-      break;
-    }
-
-    default:
-      break;
-  }
-}
-
-GfCamera CameraData::export_gf(GfVec4f tile)
-{
-  float t_pos[2] = {tile[0], tile[1]}, t_size[2] = {tile[2], tile[3]};
-
-  GfCamera gf_camera = GfCamera();
-
-  gf_camera.SetClippingRange(clip_range);
-
-  float l_shift[2] = {(lens_shift[0] + t_pos[0] + t_size[0] * 0.5f - 0.5f) / t_size[0],
-                      (lens_shift[1] + t_pos[1] + t_size[1] * 0.5f - 0.5f) / t_size[1]};
-
-  switch (mode) {
-    case CAM_PERSP:
-    case CAM_PANO: {
-      /*  TODO: store panoramic camera settings */
-      gf_camera.SetProjection(GfCamera::Projection::Perspective);
-      gf_camera.SetFocalLength(focal_length);
-
-      float s_size[2] = {sensor_size[0] * t_size[0], sensor_size[1] * t_size[1]};
-
-      gf_camera.SetHorizontalAperture(s_size[0]);
-      gf_camera.SetVerticalAperture(s_size[1]);
-
-      gf_camera.SetHorizontalApertureOffset(l_shift[0] * s_size[0]);
-      gf_camera.SetVerticalApertureOffset(l_shift[1] * s_size[1]);
-      break;
-    }
-    case CAM_ORTHO: {
-      gf_camera.SetProjection(GfCamera::Projection::Orthographic);
-
-      // Use tenths of a world unit accorging to USD docs https://graphics.pixar.com/usd/docs/api/class_gf_camera.html
-      float o_size[2] = {ortho_size[0] * t_size[0] * 10, 
-                         ortho_size[1] * t_size[1] * 10};
-
-      gf_camera.SetHorizontalAperture(o_size[0]);
-      gf_camera.SetVerticalAperture(o_size[1]);
-
-      gf_camera.SetHorizontalApertureOffset(l_shift[0] * o_size[0]);
-      gf_camera.SetVerticalApertureOffset(l_shift[1] * o_size[1]);
-      break;
-    }
-    default:
-      break;
-  }
-
-  gf_camera.SetTransform(transform);
-  
-  return gf_camera;
-}
 
 ViewSettings::ViewSettings(BL::Context &b_context)
   : camera_data(b_context)
